pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "realtime-diva"
version = "0.1.3"
authors = [
  {name="Will Held", email="wheld3@gatech.edu"},
]
description = "A Framework for User Testing Speech LLMs"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
]
dependencies = [
    "livekit>=0.9.0",
    "livekit-api",
    "livekit-agents>=0.4.0",
    "livekit-plugins-cartesia",
    "livekit-plugins-silero",
    "python-dotenv",
]

[tool.hatch.build]
include = ["src/"]
dev-mode-dirs = [".", "src"]

[tool.hatch.build.sources]
"realtime-diva" = "src/realtime"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.sdist]
only-include = ["src"]

[tool.hatch.build.targets.wheel]
only-include = ["src/realtime"]

[tool.hatch.version]
[project.urls]
"Homepage" = "https://diva-audio.github.io"

[tool.black]
line-length = 119
target-version = ["py310"]
preview = true

[tool.isort]
profile = "black"
multi_line_output = 3
lines_after_imports = 2
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 119
src_paths = ["src"]